apps/ml/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
asyncpg==0.30.0
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.1.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
